Resolve conflicts in cache_controller.py and hiradix_cache.py, preserving the
PR's lazy host-memory allocation design (allocate host_indices in
_drain_storage_hit() only after the L3 storage-hit count is confirmed) on top of
main's renamed prefetch_key / RadixKey and _all_reduce_attn_groups helpers.

+++ python/sglang/srt/lora/triton_ops/fused_moe_lora_kernel.py
@@ -186,7 +186,7 @@ def _fused_moe_lora_kernel(
             mask=token_mask[:, None] & (offs_k[None, :] < k_remaining),
             other=0.0,
         )
-        accumulator += tl.dot(a, b)
+        accumulator += tl.dot(a, b.to(a.dtype))
         # Advance the ptrs to the next K block.
         a_ptrs += BLOCK_SIZE_K * SPLIT_K * stride_ak
         b_ptrs += BLOCK_SIZE_K * SPLIT_K * stride_bk
